requirements.txt
flask
requests
python-dotenv
gunicorn
psycopg2-binary
supabase
tzdata